+++ app/views/articles/show.html.erb
@@ -204,7 +204,7 @@
       </article>
 
       <% cache("article-bottom-content-#{@article.id}-#{user_signed_in?}", expires_in: 18.hours) do %>
-        <% suggested_articles = ArticleSuggester.new(@article).articles(max: 4) %>
+        <% suggested_articles = Articles::Suggest.call(@article, max: 4) %>
         <%= render "articles/bottom_content", articles: suggested_articles %>
       <% end %>
     </div>
